Fix bug in update ranking event handler function
Created a for-loop which runs multiple patch requests to the API in
order to update the ranking of the destinations. The return promises
of the ajax requests were stored in a promises array and promise.all
was called on the promises array to subsequently display the new
destination list with the updated user ratings.

Included this logic in both the update rating event funciton and the
add destination event function so that the latest rating is always
displayed to the user.